Re-organize some API utils code
