+++ mathesar_ui/src/api/types/queries.ts
@@ -1,4 +1,4 @@
-import type { PaginatedResponse } from '@mathesar/utils/api';
+import type { PaginatedResponse } from '@mathesar/api/utils/requestUtils';
 import type { Column } from '@mathesar/api/types/tables/columns';
 import type { JpPath } from '@mathesar/api/types/tables/joinable_tables';
 import type { SchemaEntry } from '@mathesar/AppTypes';
