+++ mathesar_ui/src/stores/abstract-types/store.ts
@@ -1,5 +1,5 @@
 import { derived, writable, get } from 'svelte/store';
-import { getAPI, States } from '@mathesar/utils/api';
+import { getAPI, States } from '@mathesar/api/utils/requestUtils';
 import { currentDBId } from '@mathesar/stores/databases';
 import { preloadCommonData } from '@mathesar/utils/preloadData';
