+++ mathesar_ui/src/stores/abstract-types/types.ts
@@ -7,7 +7,7 @@ import type {
 import type { CellDataType } from '@mathesar/components/cell-fabric/data-types/typeDefinitions';
 import type { DbType } from '@mathesar/AppTypes';
 import type { Column } from '@mathesar/api/types/tables/columns';
-import type { States } from '@mathesar/utils/api';
+import type { States } from '@mathesar/api/utils/requestUtils';
 import type { abstractTypeCategory } from './constants';
 
 type AbstractTypeCategoryKeys = keyof typeof abstractTypeCategory;
