+++ mathesar_ui/src/stores/table-data/TableStructure.ts
@@ -3,7 +3,7 @@ import { derived } from 'svelte/store';
 
 import type { DBObjectEntry } from '@mathesar/AppTypes';
 import type { AbstractTypesMap } from '@mathesar/stores/abstract-types/types';
-import { States } from '@mathesar/utils/api';
+import { States } from '@mathesar/api/utils/requestUtils';
 import type { Column } from '@mathesar/api/types/tables/columns';
 import { ColumnsDataStore } from './columns';
 import type { ConstraintsData } from './constraints';
