+++ mathesar_ui/src/stores/table-data/columns.ts
@@ -6,8 +6,16 @@ import {
   WritableSet,
 } from '@mathesar-component-library';
 import type { Column } from '@mathesar/api/types/tables/columns';
-import type { PaginatedResponse, RequestStatus } from '@mathesar/utils/api';
-import { deleteAPI, getAPI, patchAPI, postAPI } from '@mathesar/utils/api';
+import type {
+  PaginatedResponse,
+  RequestStatus,
+} from '@mathesar/api/utils/requestUtils';
+import {
+  deleteAPI,
+  getAPI,
+  patchAPI,
+  postAPI,
+} from '@mathesar/api/utils/requestUtils';
 import { getErrorMessage } from '@mathesar/utils/errors';
 
 function api(url: string) {
