+++ mathesar_ui/src/stores/table-data/tabularData.ts
@@ -8,7 +8,7 @@ import {
 } from 'svelte/store';
 import type { DBObjectEntry } from '@mathesar/AppTypes';
 import type { AbstractTypesMap } from '@mathesar/stores/abstract-types/types';
-import { States } from '@mathesar/utils/api';
+import { States } from '@mathesar/api/utils/requestUtils';
 import type { Column } from '@mathesar/api/types/tables/columns';
 import { SheetSelection } from '@mathesar/components/sheet';
 import { Meta } from './meta';
