+++ mathesar_ui/src/stores/table-data/utils.ts
@@ -5,8 +5,8 @@ import {
   WritableMap,
 } from '@mathesar-component-library';
 import type { Column } from '@mathesar/api/types/tables/columns';
-import type { RequestStatus } from '@mathesar/utils/api';
-import { getMostImportantRequestStatusState } from '@mathesar/utils/api';
+import type { RequestStatus } from '@mathesar/api/utils/requestUtils';
+import { getMostImportantRequestStatusState } from '@mathesar/api/utils/requestUtils';
 import type { RowStatus } from './meta';
 import type { RecordRow } from './records';
