+++ mathesar_ui/src/systems/table-view/link-table/LinkTableModal.svelte
@@ -30,7 +30,7 @@
     getTabularDataStoreFromContext,
   } from '@mathesar/stores/table-data';
   import { toast } from '@mathesar/stores/toast';
-  import { postAPI } from '@mathesar/utils/api';
+  import { postAPI } from '@mathesar/api/utils/requestUtils';
   import { getAvailableName } from '@mathesar/utils/db';
   import { getErrorMessage } from '@mathesar/utils/errors';
   import { iconTechnicalExplanation, iconTableLink } from '@mathesar/icons';
