+++ mathesar_ui/src/utils/__tests__/api.test.ts
@@ -1,4 +1,4 @@
-import { getMostImportantRequestStatusState } from '../api';
+import { getMostImportantRequestStatusState } from '../../api/utils/requestUtils';
 
 test('getMostImportantRequestStatusState', () => {
   expect(
